chore(rspack): move apply-react-config to plugins/utils directory for consistency
