+++ packages/rspack/src/utils/model.ts
@@ -17,9 +17,3 @@ export interface EmittedFile {
   initial: boolean;
   asset?: boolean;
 }
-
-export interface SvgrOptions {
-  svgo?: boolean;
-  titleProp?: boolean;
-  ref?: boolean;
-}
